corner_case/requirements.txt
asgiref==3.3.4
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.4
djangorestframework==3.12.4
idna==2.10
importlib-metadata==4.5.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jsonpickle==2.0.0
jsonschema==3.2.0
MarkupSafe==2.0.1
packaging==20.9
pyparsing==2.4.7
pyrsistent==0.17.3
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
ruamel.yaml==0.17.9
ruamel.yaml.clib==0.2.2
six==1.16.0
sqlparse==0.4.1
swagger-spec-validator==2.7.3
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.5
zipp==3.4.1
psycopg2>=2.8.6,<2.8.7
gunicorn==20.1.0
